Tools have meaning
Delegated drawing functionality and point processing to the individual tools. This lets each tool define how their data is stored, and how their data is read.

The rectangle tool now works! The background color option disappears when you are on other tools.

Adjusted how CanvasActual is created. It is now loaded in whatever canvas-wrap is shown on the screen. The user then instantiates the CanvasActual object on their own. In the future, I may narrow this down let the user pass an element into the constructor, telling the CA where to create itself. I'd also like to design it where it is its own custom HTML5 tag as well.