TempPoint Init Edits
Changed how the currently drawn point gets its initial values. The tempPoint array now gets all of its initial values from the Tool object itself, with any Tool specific changes coming from subclasses. See the Rectangle Tool for an example.

The initPoints function also serves as an updater for whenever the current tool is changed. It grabs the value of the settings in the toolbar. This should be extensible in the future and allow for setting and changing of things like linecap and linejoin. I may want to move the width and height variables that shapes use into this object, to allow for users to set a specific width and height for the shapes they want to draw.