Bug Fixes
Fixed a HUGE bug in how colors were being stored. When I changed the variable stroke a few commits back, I neglected to change the drawing methods. Most of them were just grabbing the currently used color, not their own.

Made Eraser do something. Not much, but something. It is basically a white pen. Not sure it needs to be there at all.